treewide: Move old tests around
