+++ src/algebra/gf2/share8.rs
@@ -99,7 +99,7 @@ mod benchmark {
 
         let mut rng = thread_rng();
 
-        let mut v: [BitSharing8; 4] = [
+        let v: [BitSharing8; 4] = [
             BitSharing8(rng.gen()),
             BitSharing8(rng.gen()),
             BitSharing8(rng.gen()),
@@ -128,7 +128,7 @@ mod benchmark {
 
         b.iter(|| {
             black_box(for _ in 0..1_000_000 {
-                black_box({ v.action(s) });
+                black_box(v.action(s));
             })
         });
     }
